Added Searching.Benchmarks project and MaxSumSubArray functionality
A new project `Searching.Benchmarks` was added to the `Algorithms.sln` solution for benchmarking search algorithms. The `GlobalUsings.cs` file was updated to include global usings for `Searching.Common` and `BenchmarkDotNet.Attributes`. A new method `RunFindingMaxSumSubArray` was added to the `FindingExample` class to demonstrate finding the maximum sum subarray. The `Program.cs` file was updated to call this new method. A new class `MaxSumSubArray` was added with methods for finding the maximum sum subarray using different algorithms. The `MergeSortBenchmark.cs` file was updated to remove an unnecessary using statement. The `Searching.Benchmarks.csproj` file was added to define the new project. A new class `MaxSumSubArrayBenchmark` was added to benchmark the performance of the different maximum sum subarray algorithms. The `Program.cs` file in the `Searching.Benchmarks` project was updated to run the `MaxSumSubArrayBenchmark` using BenchmarkDotNet. Several benchmark report files were added with the results of running the `MaxSumSubArrayBenchmark`. The `MaxSumSubArrayShould.cs` file was updated with a new namespace `Searching.Tests.Common` and a new class `MaxSumSubArrayShould` containing several unit tests for the `MaxSumSubArray` class methods using the xUnit testing framework. Two tests were added to check if the methods throw an `ArgumentException` when the input array is null or empty. Several test cases were added for the `Find`, `FindBruteForce`, `FindHybrid`, and `FindKadane` methods of the `MaxSumSubArray` class using the `InlineData` attribute. The `Should` Fluent Assertions library was used to verify the test results.