Refactor RSelect to generic, add WeightedMedian class
Refactored the RSelect class to be generic, allowing it to work with different numeric types by implementing the INumber<T> interface. Updated ClosestNeighborsSearch to use the generic RSelect<T> for finding the k-th smallest element. Added a new WeightedMedian class to compute the weighted median of a set of elements and their weights. Updated RSelectShould tests to handle the tuple returned by RSelect<T>.Find and added new tests for the WeightedMedian class.