Add IntersectWith method and update related tests
A new method `IntersectWith` was added to the `GraphBase` class in `GraphBase.cs`. This method removes vertices and edges from the current graph that are not present in another graph and updates edge lengths to the minimum values if the graph has variable edge lengths.

A new test method `IntersectWith_WhenGraphsHaveCommonVertices_ShouldLeaveOnlyCommonVerticesAndEdges` was added to `UndirectedGraphShould` in `UndirectedGraphShould.cs` to verify the functionality of the `IntersectWith` method.

Renamed the existing test method `SearchCycle_OrientedGraphWithSycle_CycleFound` to `SearchCycle_WhenOrientedGraphWithSycle_CycleFound` for better clarity.

Updated the project file `Graphs.Tests.csproj` to reflect the renaming of the test file from `SerializedGraphs\SearchCycle_OrientedGraphWithSycle_CycleFound.txt` to `SerializedGraphs\SearchCycle_WhenOrientedGraphWithSycle_CycleFound.txt`.